feat:pid制御追加。目標の速度を送信できるようになった。書き方については改善の余地あり。

+++ CAN/app/include/app/robomas_peripheral.hpp
@@ -31,7 +31,7 @@
 #define motor_8 0x208
 
 struct MotorFeedback {
-    uint32_t angle;
+    uint16_t angle;
     int16_t speed;
     int16_t current;
     uint8_t temp;
